Create Likert Chart from Descriptives Data Frame

Create Likert Chart from Descriptives Data Frame

Usage

create_chart_likert(
  data,
  y = "value",
  x = "label",
  group = "variable",
  labels = "data_label",
  label_font_size = 10,
  main_font_size = 8,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  vertical = FALSE,
  what = "percent",
  seed = 1
)

Arguments

data

Data frame or tibble.

y

[character(1)]
Name of column in data for .

x

[character(1)]
Name of column in data for

group

[character(1)]
Name of column in data for

labels

[character(1)]
Name of column in data for labels.

label_font_size

[integer(1)]
Font size for data labels

main_font_size

[integer(1)]
Font size for all other text Must contain at least the number of unique values (excl. missing) in the data set.

font_family

Word font family. See officer::fp_text

colour_palette

[character()]
Must contain at least the number of unique values (incl. missing) in the data set.

colour_na

[character(1)]
Colour as a single string.

colour_2nd_binary_cat

[character(1)]
Colour for second category in binary variables. Often useful to hide this.

vertical

Logical. If FALSE (default), then horizontal.

what

[character(1)] Either "percent" or "frequency". Supports partial matching.

seed

Optional random seed for selection of colours in blender.

Value

mschart-object. Can be added to an rdocx, rpptx or rxlsx object.

Examples

#create_chart_likert(prepare_data_for_mschart(ex_survey1[paste0("b_", 1:3)]))