Title

Title

Usage

crosstable_to_apa(
  x,
  label_separator = NULL,
  docx_template = NULL,
  caption_style = NULL,
  body_style = NULL,
  table_header_style = NULL,
  footer_style = NULL,
  caption_autonum = NULL,
  topcaption = TRUE,
  return_docx = TRUE
)

Arguments

x

Crosstable

label_separator

[character(1)]
If not NULL (default), will split labels into main- and sub-questions and create figure caption.

docx_template

[character(1) || officer::read_docx()]

caption_style

[character(1)]
Word template style to be used for formatting chart. Defaults to "Normal".

body_style

[character(1)]
Word style for table body cells.

table_header_style

[character(1)]
Word style for table header cells.

footer_style

[character(1)]
Word style for table footer.

caption_autonum

Object obtained from run_autonum.

topcaption

[logical(1)]
Place caption above (TRUE, default) or below table.

return_docx

[logical(1)]
If FALSE, will return the table rather than a rdocx object with the table inside it. Set to TRUE (default) if piping multiple tables and charts together. Set to FALSE if you want to continue modify the table.

Value

rdocx object, which can be saved with print() after loading the officer-package

Examples

library(crosstable)
library(dplyr)
library(officer)
crosstable::crosstable(ex_survey1, b_1:b_3, percent_pattern = "{p_col}") %>%
  crosstable_to_apa(label_separator=" - ")
#> rdocx document with 4 element(s)
#> 
#> * styles:
#>                 Normal              heading 1              heading 2 
#>            "paragraph"            "paragraph"            "paragraph" 
#>              heading 3 Default Paragraph Font           Normal Table 
#>            "paragraph"            "character"                "table" 
#>                No List                 strong               centered 
#>            "numbering"            "character"            "paragraph" 
#>         table_template    Light List Accent 2            Titre 1 Car 
#>                "table"                "table"            "character" 
#>            Titre 2 Car            Titre 3 Car          Image Caption 
#>            "character"            "character"            "paragraph" 
#>          Table Caption     Table Professional                  toc 1 
#>            "paragraph"                "table"            "paragraph" 
#>                  toc 2           Balloon Text    Texte de bulles Car 
#>            "paragraph"            "paragraph"            "character" 
#>           reference_id          graphic title            table title 
#>            "character"            "paragraph"            "paragraph" 
#> 
#> * Content at cursor location:
#>   level num_id     text style_name content_type
#> 1    NA     NA Note. N=     Normal    paragraph