Create Word Report with Univariates for Categorical Columns Sharing Same Categories

Create Word Report with Univariates for Categorical Columns Sharing Same Categories

Usage

report_chart_likert(
  data,
  cols = everything(),
  by = NULL,
  showNA = "ifany",
  docx_template = NULL,
  label_font_size = 8,
  main_font_size = 9,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  height_per_col = 0.3,
  height_fixed = 1,
  chart_formatting = NULL,
  caption_style = NULL,
  what = "percent",
  digits = 1,
  percent_sign = TRUE,
  sort_by = NULL,
  vertical = FALSE,
  desc = FALSE,
  hide_label_if_below = 1,
  label_separator = NULL,
  caption_autonum = NULL,
  seed = 1
)

Arguments

data

Data frame or tibble.

cols

Columns to select for reporting. Supports tidy-select.

by

data-masking
. Optional column used to break graph by.

showNA

[logical(1)>0]
Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).

docx_template

[character(1) || officer::read_docx()]
Either a filepath to a template file, or a rdocx-object from read_docx.

label_font_size

[integer(1)]
Font size for data labels

main_font_size

[integer(1)]
Font size for all other text

font_family

[character(1)]
Office font family. Defaults to "Arial". See ?officer::fp_text() for options.

colour_palette

[character()]
Must contain at least the number of unique values (incl. missing) in the data set.

colour_na

[character(1)]
Colour as a single string, for NA-values.

colour_2nd_binary_cat

[character(1)]
Colour for second category in binary variables. Often useful to hide this.

height_per_col

[numeric(1)>0]
Height in cm per chart entry.

height_fixed

[numeric(1)>0]
Fixed height in cm.

chart_formatting

[character(1)]
Template style to be used for formatting chart

caption_style

[character(1)]
Template style to be used for formatting chart. Defaults to "Normal".

what

[character(1)] Either "percent" or "frequency". Supports partial matching.

digits

[integer(1)]
Number of decimal places as integer.

percent_sign

[logical(1)>0]
Whether to include percentage symbol on chart.

sort_by

[character(1)], sort output by "label", size of a category, or sum of categories (character vector). Defaults to none (NULL).

vertical

[logical(1)>0]
If FALSE (default), then horizontal.

desc

[logical(1)>0]
Sort in descending order?

hide_label_if_below

[numeric(1)]
Whether to hide label if below this value.

label_separator

[character(1)]
If not NULL (default), will split labels into main- and sub-questions and create figure caption.

caption_autonum

Object obtained from run_autonum.

seed

[numeric(1)>0]
Optional random seed for selection of colours in blender.

Value

rdocx object, which can be saved with officer:::print.rdocx()

Examples

library(dplyr) # For piping
library(officer) # To save the rdocx object to disk
ex_survey1 %>%
  report_chart_likert(cols = a_1:a_9) %>%
  print(target = "test_docx_a19.docx")
file.remove("test_docx_a19.docx")
#> [1] TRUE

  docx_template <-
      system.file("template","NIFUmal_tom.docx",
                   package = "surveyreport", mustWork = TRUE)
  colour_palette <-
    readxl::read_excel(system.file("template", "NIFUmal_stiler.xlsx",
                                   package = "surveyreport", mustWork = TRUE),
                       sheet = "NIFUblue") %>%
    dplyr::pull(hex)
chart_format <-
 system.file("template", "NIFUmal_stiler.xlsx",
             package = "surveyreport", mustWork = TRUE) %>%
 readxl::read_excel(., sheet = 1) %>%
 dplyr::filter(surveyreport_style == "figure") %>%
 dplyr::pull(template_style)

 test_docx_b13 <-
   ex_survey1 %>%
   report_chart_likert(cols = b_1:b_3,
                       docx_template = docx_template,
                       colour_palette = colour_palette,
                       chart_formatting = chart_format,
                       height_per_col = .3,
                       height_fixed = 1)
print(test_docx_b13, target = "test_docx_b13.docx")
file.remove("test_docx_b13.docx")
#> [1] TRUE