Extention to the infer package for multiple testing

Extention to the infer package for multiple testing

Usage

test_multiple_comb(
  x,
  formula = NULL,
  response = NULL,
  explanatory = NULL,
  order = NULL,
  alternative = "two-sided",
  mu = 0,
  correct = TRUE,
  p = NULL,
  conf_int = TRUE,
  conf_level = 0.95,
  test = c("t", "chisq", "prop"),
  success = NULL,
  z = FALSE
)

Arguments

x

A data frame that can be coerced into a tibble.

formula

A formula with the response variable on the left and the explanatory on the right.

response

The variable name in x that will serve as the response. This is alternative to using the formula argument.

explanatory

The variable name in x that will serve as the explanatory variable.

order

A string vector of specifying the order in which the levels of the explanatory variable should be ordered for subtraction, where order = c("first", "second") means ("first" - "second").

alternative

Character string giving the direction of the alternative hypothesis. Options are "two-sided" (default), "greater", or "less".

mu

A numeric value giving the hypothesized null mean value for a one sample test and the hypothesized difference for a two sample test.

correct

A logical indicating whether Yates' continuity correction should be applied where possible. If z = TRUE, the correct argument will be overwritten as FALSE. Otherwise defaults to correct = TRUE.

p

A numeric vector giving the hypothesized null proportion of success for each group.

conf_int

A logical value for whether to include the confidence interval or not. TRUE by default.

conf_level

A numeric value between 0 and 1. Default value is 0.95.

test

String indicating type of test done for all combinations.

success

The level of response that will be considered a success, as a string. Only used when testing the null that a single proportion equals a given value, or that two proportions are equal; ignored otherwise.

z

A logical value for whether to report the statistic as a standard normal deviate or a Pearson's chi-square statistic. \(z^2\) is distributed chi-square with 1 degree of freedom, though note that the user will likely need to turn off Yates' continuity correction by setting correct = FALSE to see this connection.

Value

A tibble containing for each response-explanatory row combination columns summarizing the test statistic.

Examples

library(dplyr)
x <- test_multiple_comb(x = mtcars,
                     response = matches("mpg|disp|drat|hp|qsec"),
                     explanatory = matches("vs|am"),
                     test = "t")
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.
#> Warning: The statistic is based on a difference or ratio; by default, for difference-based statistics, the explanatory variable is subtracted in the order "0" - "1", or divided in the order "0" / "1" for ratio-based statistics. To specify this order yourself, supply `order = c("0", "1")`.